% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shp.reader.R
\name{dbf.reader}
\alias{dbf.reader}
\title{Overriden dbf reader}
\usage{
dbf.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\description{
This alternative reader avoids loading shapefiles' dbf companion files as if
they were XBASE files
}

