% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shp.reader.R
\name{shp.reader}
\alias{shp.reader}
\title{Read a shapefile described in a .shp file.}
\usage{
shp.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load data from an ElasticSearch index based on configuration
information found in the specified .es file. The .es file must specify
an ElasticSearch sercer to be accessed. A specific query against any index may be executed to generate
a data set.
}
\examples{
library('ProjectTemplate')

\dontrun{es.reader('example.es', 'data/example.es', 'example')}
}

